requirements.txt
pandas
matplotlib
numpy
streamlitgit